+++ app/javascript/packs/application.ts
