+++ app/javascript/src/components/Forecast/Forecast.tsx
