+++ app/javascript/src/components/Forecast/ForecastListItem.tsx
