+++ app/javascript/src/components/Forecast/ForecastSelectorDate.tsx
