+++ package.json
@@ -4,12 +4,15 @@
   "dependencies": {
     "@apollo/react-hooks": "^4.0.0",
     "@babel/preset-react": "^7.12.10",
+    "@babel/preset-typescript": "^7.12.16",
     "@fortawesome/fontawesome-free": "^5.15.2",
     "@rails/actioncable": "^6.0.0",
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
     "@rails/webpacker": "5.2.1",
     "@reach/router": "^1.3.4",
+    "@types/react": "^17.0.2",
+    "@types/react-dom": "^17.0.1",
     "apollo-boost": "^0.4.9",
     "apollo-cache-inmemory": "^1.6.6",
     "apollo-client": "^2.6.10",
@@ -26,7 +29,8 @@
     "react": "^17.0.1",
     "react-apollo": "^3.1.5",
     "react-dom": "^17.0.1",
-    "turbolinks": "^5.2.0"
+    "turbolinks": "^5.2.0",
+    "typescript": "^4.1.5"
   },
   "scripts": {
     "lint": "eslint --ext .js --ext .jsx app/javascript/",
